changes, what went wrong

+++ Car_Controller/LineSensor.cpp
@@ -29,7 +29,7 @@ int LineSensor::readRightState()
 }
 
 
-int LineSensor::readState()
+byte LineSensor::readState()
 {
   leftState = analogRead(_leftPin);
   rightState = analogRead(_rightPin);
